+++ lib/webdrivers.rb
@@ -1,2 +1,4 @@
 require "webdrivers/selenium"
+require "webdrivers/common"
 require "webdrivers/chromedriver"
+require "webdrivers/geckodriver"
